added @loginrequired to perform create, update, delete, to see profile and order history

+++ catalogue/views.py
@@ -4,6 +4,7 @@ from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
 from .forms import ProductForm
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
@@ -72,8 +73,13 @@ def product_description(request, product_id):
     return render(request, 'catalogue/product_description.html', context)
 
 
+@login_required
 def create_product(request):
     """ Creates a new product """
+    if not request.user.is_superuser:
+        messages.error(request, 'Whoops, only store owners here')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -93,8 +99,13 @@ def create_product(request):
     return render(request, template, context)
 
 
+@login_required
 def update_product(request, product_id):
     """ Edits an existing product """
+    if not request.user.is_superuser:
+        messages.error(request, 'Whoops, only store owners here')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -117,8 +128,13 @@ def update_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Deletes an existing product """
+    if not request.user.is_superuser:
+        messages.error(request, 'Whoops, only store owners here')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, f'Product {product.name} has been deleted')
